reworded suffix and prefix, fixed if no valid steps, and moved coloring, formating, progress bar test of model in their own file
