+++ tests/framework/model/test_model.py
@@ -25,11 +25,6 @@ try:
 except ImportError:
     MNIST = None
 
-try:
-    import colorama as color
-except ImportError:
-    color = None
-
 warning_settings['concatenate_returns'] = 'ignore'
 
 
@@ -148,225 +143,6 @@ class ModelTest(ModelFittingTestCase):
             'valid_steps': ModelTest.steps_per_epoch
         }
         self._test_callbacks_train(params, logs, valid_steps=ModelTest.steps_per_epoch)
-
-    def assertStdoutContains(self, values):
-        for value in values:
-            self.assertIn(value, self.test_out.getvalue().strip())
-
-    def assertStdoutNotContains(self, values):
-        for value in values:
-            self.assertNotIn(value, self.test_out.getvalue().strip())
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_fitting_with_default_coloring(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback])
-
-        self.assertStdoutContains(["[32m", "[35m", "[36m", "[94m"])
-
-    def test_fitting_with_progress_bar_show_epoch(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback])
-
-        self.assertStdoutContains(["Epoch", "1/5", "2/5"])
-
-    def test_fitting_with_no_progress_bar__dont_show_epoch(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     verbose=False)
-
-        self.assertStdoutNotContains(["Epoch", "1/5", "2/5"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_fitting_with_user_coloring(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        coloring = {
-            "text_color": 'BLACK',
-            "ratio_color": "BLACK",
-            "metric_value_color": "BLACK",
-            "time_color": "BLACK",
-            "progress_bar_color": "BLACK"
-        }
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=coloring))
-
-        self.assertStdoutContains(["[30m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_fitting_with_user_partial_coloring(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring={
-                                         "text_color": 'BLACK',
-                                         "ratio_color": "BLACK"
-                                     }))
-
-        self.assertStdoutContains(["[30m", "[32m", "[35m", "[94m"])
-
-    def test_fitting_with_user_coloring_invalid(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        with self.assertRaises(KeyError):
-            _ = self.model.fit_generator(train_generator,
-                                         valid_generator,
-                                         epochs=ModelTest.epochs,
-                                         steps_per_epoch=ModelTest.steps_per_epoch,
-                                         validation_steps=ModelTest.steps_per_epoch,
-                                         callbacks=[self.mock_callback],
-                                         progress_options=dict(coloring={"invalid_name": 'A COLOR'}))
-
-    def test_fitting_with_no_coloring(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=False))
-
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_fitting_with_progress_bar_default_color(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=True, progress_bar=True))
-
-        self.assertStdoutContains(["%", "[32m", "[35m", "[36m", "[94m", "\u2588"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_fitting_with_progress_bar_user_color(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        coloring = {
-            "text_color": 'BLACK',
-            "ratio_color": "BLACK",
-            "metric_value_color": "BLACK",
-            "time_color": "BLACK",
-            "progress_bar_color": "BLACK"
-        }
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=coloring, progress_bar=True))
-
-        self.assertStdoutContains(["%", "[30m", "\u2588"])
-
-    def test_fitting_with_progress_bar_no_color(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=False, progress_bar=True))
-
-        self.assertStdoutContains(["%", "\u2588"])
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    def test_fitting_with_no_progress_bar(self):
-        train_generator = some_data_tensor_generator(ModelTest.batch_size)
-        valid_generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     steps_per_epoch=ModelTest.steps_per_epoch,
-                                     validation_steps=ModelTest.steps_per_epoch,
-                                     callbacks=[self.mock_callback],
-                                     progress_options=dict(coloring=False, progress_bar=False))
-
-        self.assertStdoutNotContains(["%", "\u2588"])
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    def test_progress_bar_with_step_is_none(self):
-        train_generator = SomeDataGeneratorUsingStopIteration(ModelTest.batch_size, 10)
-        valid_generator = SomeDataGeneratorUsingStopIteration(ModelTest.batch_size, 10)
-
-        self._capture_output()
-
-        _ = self.model.fit_generator(train_generator,
-                                     valid_generator,
-                                     epochs=ModelTest.epochs,
-                                     progress_options=dict(coloring=False, progress_bar=True))
-
-        self.assertStdoutContains(["s/step"])
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m", "\u2588", "%"])
 
     def test_fitting_without_valid_generator(self):
         train_generator = some_data_tensor_generator(ModelTest.batch_size)
@@ -737,147 +513,6 @@ class ModelTest(ModelFittingTestCase):
         self.assertEqual(type(metrics), np.ndarray)
         self.assertEqual(metrics.tolist(), self.batch_metrics_values + self.epoch_metrics_values)
 
-    def test_evaluate_without_progress_output(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x, y, batch_size=ModelTest.batch_size, verbose=False)
-
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_default_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x, y, batch_size=ModelTest.batch_size)
-
-        self.assertStdoutContains(["[32m", "[35m", "[36m", "[94m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_user_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        coloring = {
-            "text_color": 'BLACK',
-            "ratio_color": "BLACK",
-            "metric_value_color": "BLACK",
-            "time_color": "BLACK",
-            "progress_bar_color": "BLACK"
-        }
-
-        _, _ = self.model.evaluate(x, y, batch_size=ModelTest.batch_size, progress_options=dict(coloring=coloring))
-
-        self.assertStdoutContains(["[30m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_user_partial_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x,
-                                   y,
-                                   batch_size=ModelTest.batch_size,
-                                   progress_options=dict(coloring={
-                                       "text_color": 'BLACK',
-                                       "ratio_color": "BLACK"
-                                   }))
-        self.assertStdoutContains(["[30m", "[32m", "[35m", "[94m"])
-
-    def test_evaluate_with_user_coloring_invalid(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        with self.assertRaises(KeyError):
-            _, _ = self.model.evaluate(x,
-                                       y,
-                                       batch_size=ModelTest.batch_size,
-                                       progress_options=dict(coloring={"invalid_name": 'A COLOR'}))
-
-    def test_evaluate_with_no_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x, y, batch_size=ModelTest.batch_size, progress_options=dict(coloring=False))
-
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_progress_bar_default_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x,
-                                   y,
-                                   batch_size=ModelTest.batch_size,
-                                   progress_options=dict(coloring=True, progress_bar=True))
-
-        self.assertStdoutContains(["%", "[32m", "[35m", "[36m", "[94m", "\u2588"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_progress_bar_user_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        coloring = {
-            "text_color": 'BLACK',
-            "ratio_color": "BLACK",
-            "metric_value_color": "BLACK",
-            "time_color": "BLACK",
-            "progress_bar_color": "BLACK"
-        }
-
-        _, _ = self.model.evaluate(x,
-                                   y,
-                                   batch_size=ModelTest.batch_size,
-                                   progress_options=dict(coloring=coloring, progress_bar=True))
-
-        self.assertStdoutContains(["%", "[30m", "\u2588"])
-
-    @skipIf(color is None, "Unable to import colorama")
-    def test_evaluate_with_progress_bar_user_no_color(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x,
-                                   y,
-                                   batch_size=ModelTest.batch_size,
-                                   progress_options=dict(coloring=False, progress_bar=True))
-
-        self.assertStdoutContains(["%", "\u2588"])
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
-    def test_evaluate_with_no_progress_bar(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate(x,
-                                   y,
-                                   batch_size=ModelTest.batch_size,
-                                   progress_options=dict(coloring=False, progress_bar=False))
-
-        self.assertStdoutNotContains(["%", "\u2588"])
-        self.assertStdoutNotContains(["[32m", "[35m", "[36m", "[94m"])
-
     def test_evaluate_with_pred(self):
         x = torch.rand(ModelTest.evaluate_dataset_len, 1)
         y = torch.rand(ModelTest.evaluate_dataset_len, 1)
@@ -923,18 +558,6 @@ class ModelTest(ModelFittingTestCase):
         self.assertEqual(metrics.tolist(), self.batch_metrics_values + self.epoch_metrics_values)
         self.assertEqual(pred_y.shape, (ModelTest.evaluate_dataset_len, 1))
 
-    def test_evaluate_data_loader_with_progress_bar_coloring(self):
-        x = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        y = torch.rand(ModelTest.evaluate_dataset_len, 1)
-        dataset = TensorDataset(x, y)
-        generator = DataLoader(dataset, ModelTest.batch_size)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate_generator(generator, verbose=True)
-
-        self.assertStdoutContains(["%", "[32m", "[35m", "[36m", "[94m", "\u2588"])
-
     def test_evaluate_generator(self):
         num_steps = 10
         generator = some_data_tensor_generator(ModelTest.batch_size)
@@ -944,16 +567,6 @@ class ModelTest(ModelFittingTestCase):
         self.assertEqual(metrics.tolist(), self.batch_metrics_values + self.epoch_metrics_values)
         self.assertEqual(type(pred_y), np.ndarray)
         self.assertEqual(pred_y.shape, (num_steps * ModelTest.batch_size, 1))
-
-    def test_evaluate_generator_with_progress_bar_coloring(self):
-        num_steps = 10
-        generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate_generator(generator, steps=num_steps, verbose=True)
-
-        self.assertStdoutContains(["%", "[32m", "[35m", "[36m", "[94m", "\u2588"])
 
     def test_evaluate_generator_with_callback(self):
         num_steps = 10
@@ -969,16 +582,6 @@ class ModelTest(ModelFittingTestCase):
         logs = self.model.evaluate_generator(generator, steps=num_steps, return_dict_format=True)
 
         self._test_return_dict_logs(logs)
-
-    def test_evaluate_generator_with_callback_and_progress_bar_coloring(self):
-        num_steps = 10
-        generator = some_data_tensor_generator(ModelTest.batch_size)
-
-        self._capture_output()
-
-        _, _ = self.model.evaluate_generator(generator, steps=num_steps, callbacks=[self.mock_callback], verbose=True)
-
-        self.assertStdoutContains(["%", "[32m", "[35m", "[36m", "[94m", "\u2588"])
 
     def test_evaluate_generator_with_ground_truth(self):
         num_steps = 10
